+++ components/header/index.jsx
@@ -9,10 +9,10 @@ import styled from 'styled-components';
 
 import { currentUserInfo } from '../../store/states';
 import { logOut } from '../../service/auth';
-import { HeaderButton } from '../shared/Button1';
-import ProfileIcon from '../shared/ProfileIcon1';
-import MindMapInfo from '../MindMapInfo1';
-import { AUTH_EXPIRED } from '../../constants/constants1';
+import { HeaderButton } from '../shared/button';
+import ProfileIcon from '../shared/profileicon';
+import MindMapInfo from '../mindmapinfo';
+import { AUTH_EXPIRED } from '../../constants/constants';
 
 export default function Header() {
   const [mounted, setMounted] = useState(false);
