+++ components/navbar/index.jsx
@@ -5,9 +5,9 @@ import { useRouter } from 'next/router';
 import styled from 'styled-components';
 
 import { useSetRecoilState, useRecoilValue } from 'recoil';
-import { NavBarButton } from '../shared/Button1';
+import { NavBarButton } from '../shared/button';
 
-import { createMindMapData } from '../../service/mindMapRequests1';
+import { createMindMapData } from '../../service/mindmaprequests';
 import {
   currentUserInfo,
   errorInfo,
