+++ components/nodecomment/index.jsx
@@ -15,10 +15,10 @@ import {
   socketInfo,
 } from '../../store/states';
 import flexCenter from '../shared/FlexCenterContainer1';
-import { Button } from '../shared/Button1';
+import { Button } from '../shared/button';
 import { postCommentsData } from '../../service/nodeRequests1';
-import ProfileIcon from '../shared/ProfileIcon1';
-import { NO_PERMISSION_MESSAGE } from '../../constants/constants1';
+import ProfileIcon from '../shared/profileicon';
+import { NO_PERMISSION_MESSAGE } from '../../constants/constants';
 
 export default function NodeComment() {
   const commentList = useRef();
