+++ constants/nodecolor.js
