+++ pages/404.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import NotFoundPage from '../components/Error1/404';
+import NotFoundPage from '../components/error/404';
 
 export default function LoginPage() {
   return <NotFoundPage />;
