+++ pages/500.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import InternalError from '../components/Error1/500';
+import InternalError from '../components/error/500';
 
 export default function LoginPage() {
   return <InternalError />;
