+++ utils/d3/calculatenewnodeposition.js
