+++ utils/deletenodehelper.js
